+++ darjeeling-android-testing/src/main/java/no/agens/darjeeling/android/DarjeelingFragmentTest.kt
@@ -5,16 +5,25 @@ import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.testing.FragmentScenario
 import org.junit.Before
 
-abstract class DarjeelingFragmentTest<F : Fragment> : DarjeelingBase() {
+abstract class DarjeelingFragmentTest<F : Fragment>(private val autoLaunchFragments: Boolean = true) :
+    DarjeelingBase() {
 
     private lateinit var underlyingActivity: FragmentActivity
-    lateinit var scenario: FragmentScenario<F>
+    private lateinit var scenario: FragmentScenario<F>
 
     override val activity: FragmentActivity
         get() = underlyingActivity
 
     @Before
     open fun setup() {
+        if (autoLaunchFragments) {
+            launchFragment()
+        }
+    }
+
+    abstract fun createFragmentScenario(): FragmentScenario<F>
+
+    fun launchFragment() {
         scenario = createFragmentScenario()
         scenario.onFragment {
 
@@ -22,7 +31,9 @@ abstract class DarjeelingFragmentTest<F : Fragment> : DarjeelingBase() {
         }
     }
 
-    abstract fun createFragmentScenario(): FragmentScenario<F>
+    fun onFragment(action: (F) -> Unit) {
+        scenario.onFragment(action)
+    }
 
 
 }